refactor(daemon): split mcp/ops.ts tool dispatch into a per-domain handler registry
`executeTool` was an ~820-line if/else chain over tool names and `OpsDeps` a
~67-field God interface. Both split along the lines the interface groups and the
`*_TOOLS` arrays already drew.

- `src/mcp/ops/` now holds one file per domain — session, messaging, directory,
  knowledge, orchestration, github, memory, platform-reads — each exporting its
  handlers verbatim plus the deps interface its tools need. Shared pieces move to
  `context.ts` (session/gateway types), `gateway.ts` (integration resolution) and
  `validate.ts` (the hand-written validators, unchanged — zod is a follow-up).
- `executeTool` keeps the guards (canRun, the evaluation-tool hook, the memory
  access gate) and dispatches through a name→handler Map, then falls through to
  the session-gateway gate, the attachment-read predicate, and the unchanged
  unknown-tool error, in exactly the old order.
- `OpsDeps` is now the composition of the domain deps interfaces; its members are
  identical, so the daemon's wiring literal is untouched.

Behavior is byte-identical: same error messages, same per-tool validation order,
same result shapes.
