+++ packages/daemon/src/mcp/ops/context.ts
@@ -1,0 +1,106 @@
+import type { ToolDescriptor } from '../tool-descriptor.js'
+
+/**
+ * The platform-neutral slice a session's tools need. `SlackConnection` and
+ * `TelegramConnection` both implement this; tests pass a fake. Read methods
+ * return already-shaped plain objects so the tool results are stable regardless
+ * of the underlying platform SDK response. "channel" is the platform conversation
+ * id (Slack channel `C…` / Telegram chat id); "user" is a platform user id.
+ */
+/** Optional per-message sender identity for {@link MessageGateway}. Slack can render
+ *  the name/avatar (via `chat:write.customize`) and persist the stable AgentConnect
+ *  author id in message metadata; other platforms ignore these fields. */
+export interface SendIdentity {
+  username?: string
+  /** Public https image URL for the message avatar (the agent's icon). */
+  icon_url?: string
+  /** Stable AgentConnect author id persisted in Slack message metadata. */
+  agentAuthorId?: string
+  /**
+   * The finalized-response block for the VISIBLE half of a paired `toAgent + channel`
+   * send (send-message-routing-rework.md §3.2/§4).
+   *
+   * Unlike a streamed turn reply, this post is complete the moment it is made — there is
+   * no later finalization edit to close it — so it is stamped `final` here, carrying the
+   * pairing id the target's activation rendezvous keys on. Without it the platform half
+   * of a paired call arrives at ingress indistinguishable from an ordinary agent reply,
+   * and the platform-first arrival order can never be recognized as one.
+   */
+  response?: {
+    responseId: string
+    deliveryState: 'streaming' | 'final'
+    hopCount: number
+    mentionedAgentIds: string[]
+    agentCallDeliveryId?: string
+  }
+}
+
+export interface MessageGateway {
+  /** Layer-1 `openDirectMessage`: resolve one platform user to the app's real
+   *  direct-message conversation. Optional because it is a declared read port,
+   *  not a universal one — Slack needs it before posting with a customized agent
+   *  identity (sending to a raw U… id routes the message through Slack's
+   *  system-notification DM); a platform without the port has no `toUser` form. */
+  openDirectMessage?(user: string): Promise<string>
+  /** Post a message; returns the resulting message id (`ts` / message_id) so the
+   *  daemon can record it. `identity` carries the agent's stable id and optional
+   *  visual identity; other platforms may ignore it. */
+  postMessage(channel: string, text: string, threadTs?: string, identity?: SendIdentity): Promise<string | undefined>
+  /** Materialize a provider-native thread from a root message when that platform
+   *  requires one before the post can own a follow-up session (Discord). */
+  createThread?(channel: string, messageId: string, name: string): Promise<string | undefined>
+  getChannelInfo(channel: string): Promise<{ id: string; name?: string; isIm?: boolean; isPrivate?: boolean }>
+  listMembers(channel: string): Promise<{ id: string; name?: string; isBot?: boolean }[]>
+  listChannels(): Promise<{ id: string; name?: string; isPrivate?: boolean }[]>
+  getUserProfile(user: string): Promise<{ id: string; name?: string; realName?: string; isBot?: boolean }>
+  /** Download an auth-gated file (Slack url_private / Telegram file_id) with the
+   *  bot credentials; null on failure / over-cap. Backs the `read*File` tools so
+   *  the agent can read attachments without ever holding the token. */
+  downloadFile(sourceUrl: string, maxBytes?: number): Promise<Buffer | null>
+}
+
+/**
+ * A tool result that should be handed to the agent as native MCP content blocks
+ * (e.g. a viewable image) rather than JSON-stringified text. The bridge detects
+ * the `mcpContent` marker and passes the blocks through verbatim.
+ */
+export interface McpContentResult {
+  mcpContent: ({ type: 'text'; text: string } | { type: 'image'; data: string; mimeType: string })[]
+  /** Preserve a remote MCP tool's semantic error without converting its content. */
+  mcpIsError?: boolean
+}
+
+/** Per-session binding captured at `session/new`, keyed by the IPC token. */
+export interface SessionContext {
+  agentId: string
+  /** The platform that triggered this session (slack/telegram/…). Trusted — set
+   *  by the daemon from the real triggering message, never a tool input. Used to
+   *  build the coords for peer-discovery (`listAgents`). */
+  platform: string
+  /** The exact platform integration that delivered this session. Absent for a
+   *  platform-free session; daemon-local universal tools still work there. */
+  integrationId?: string
+  /** Opaque physical-bot identity captured by the daemon. Platform coordinates
+   *  can overlap across bots, so daemon-local session lookups include this scope. */
+  transportScope?: string
+  /** Whether the trusted source conversation is a direct message. Slack native
+   *  thread titles are valid only for app-DM sessions. */
+  isDm: boolean
+  channel: string
+  thread: string
+  tools: ToolDescriptor[]
+  /** Snapshot of the agent's integrations (id + platform) at session/new. Lets the
+   *  platform-neutral `sendMessage` tool route to ANY connected platform,
+   *  not just the one that triggered this session. Absent ⇒ fall back to the
+   *  current session's own integration (same-platform send only). */
+  integrations?: { id: string; platform: string }[]
+  /** The agent's display identity (displayName || name), stamped on tool sends so
+   *  they match the agent's ordinary replies. The model never supplies this. */
+  agentName?: string
+  /** The agent's public avatar URL, sibling of {@link agentName}. */
+  iconUrl?: string
+}
+
+/** One tool handler in the pre-gateway registry: the trusted session context, the raw
+ *  tool args, and the deps slice its domain declares. */
+export type ToolHandler<D> = (ctx: SessionContext, args: Record<string, unknown>, deps: D) => Promise<unknown> | unknown
