+++ packages/daemon/src/mcp/ops/memory.ts
@@ -1,0 +1,214 @@
+import { randomUUID } from 'node:crypto'
+import type { SessionContext } from './context.js'
+import {
+  optionalBoundedInt,
+  optionalObject,
+  optionalString,
+  requireString,
+  requireStringAllowEmpty
+} from './validate.js'
+import type { MemoryProvider, MemoryScope } from '../../memory/provider.js'
+import { MemoryPathError, MemoryTooLargeError } from '../../memory/store.js'
+
+/** The memory-tool deps. The access gate itself is enforced pre-dispatch in `executeTool`. */
+export interface MemoryOpsDeps {
+  /** The agent memory provider — backs the `readMemory`/`writeMemory` tools.
+   *  Universal (every agent has memory), independent of the platform. */
+  memory: MemoryProvider
+  /** Session-isolation gate for the explicit memory-tool path, by operation
+   *  (#653). Agent memory is shared across users: every session may READ it
+   *  (read/search/get), but only a non-isolated session may WRITE it
+   *  (write/save/update/delete), so a private DM/A2A turn cannot push its content
+   *  into shared memory. Automatic recall is always allowed; post-turn capture and
+   *  Dream selection are gated at their own boundaries. Checked at CALL time so a
+   *  mid-session policy change takes effect immediately. Absent ⇒ allowed (e.g. in
+   *  unit fixtures). */
+  memoryAccessAllowed?: (ctx: SessionContext, mode: 'read' | 'write') => boolean | Promise<boolean>
+  /** Build the memory scope for a tool call — carries the per-channel folder key
+   *  for a channel-scoped agent so tools read/write that channel's memory (#653).
+   *  Absent ⇒ agent-level store (unit fixtures). */
+  memoryScope?: (ctx: SessionContext) => MemoryScope
+}
+
+/** Every memory tool's access mode, checked before dispatch: reads are universal, writes
+ *  are refused for an isolated session (#653). */
+export const MEMORY_TOOL_ACCESS_MODES: Record<string, 'read' | 'write'> = {
+  readMemory: 'read',
+  writeMemory: 'write',
+  searchMemory: 'read',
+  getMemory: 'read',
+  saveMemory: 'write',
+  updateMemory: 'write',
+  deleteMemory: 'write'
+}
+
+function memoryScopeFor(ctx: SessionContext, deps: MemoryOpsDeps): MemoryScope {
+  return deps.memoryScope?.(ctx) ?? { agentId: ctx.agentId }
+}
+
+/** Map the store's typed failures onto the tool-facing messages, verbatim for both file tools. */
+function toToolError(err: unknown): never {
+  if (err instanceof MemoryPathError) throw new Error(`invalid memory path: ${err.message}`)
+  if (err instanceof MemoryTooLargeError) throw new Error(err.message)
+  throw err
+}
+
+// Memory tools are universal (every agent has memory) and daemon-local — dispatched
+// before the platform-gateway gate so an agent with no platform integration works.
+export async function readMemory(
+  ctx: SessionContext,
+  args: Record<string, unknown>,
+  deps: MemoryOpsDeps
+): Promise<unknown> {
+  const scope = memoryScopeFor(ctx, deps)
+  try {
+    const path = optionalString(args, 'path') ?? 'MEMORY.md'
+    return await deps.memory.read(scope, path)
+  } catch (err) {
+    toToolError(err)
+  }
+}
+
+export async function writeMemory(
+  ctx: SessionContext,
+  args: Record<string, unknown>,
+  deps: MemoryOpsDeps
+): Promise<unknown> {
+  const scope = memoryScopeFor(ctx, deps)
+  try {
+    // writeMemory — exactly one of two modes: full-write (`content`) OR targeted edit
+    // (`oldString` + `newString`). Validate the pair ATOMICALLY: either edit field present
+    // selects edit mode, and BOTH are then required (so a stray `newString` isn't silently
+    // ignored, and an omitted `newString` isn't silently treated as a deletion — deletion
+    // must be an explicit `newString: ""`).
+    const path = optionalString(args, 'path') ?? 'MEMORY.md'
+    const oldString = optionalString(args, 'oldString')
+    const newString = optionalString(args, 'newString')
+    const content = optionalString(args, 'content')
+    const editMode = oldString !== undefined || newString !== undefined
+    if (editMode) {
+      if (content !== undefined)
+        throw new Error('writeMemory: pass EITHER `content` (full write) OR `oldString`+`newString` (edit), not both')
+      if (oldString === undefined || newString === undefined)
+        throw new Error(
+          'writeMemory: an edit needs BOTH `oldString` and `newString` (pass `newString: ""` to delete the matched text)'
+        )
+      // str-replace: read → replace the single exact occurrence → write the whole file back.
+      // Writes are serialized per agent turn, so a read-modify-write race is not a concern.
+      const current = (await deps.memory.read(scope, path)).content
+      const occurrences = oldString === '' ? 0 : current.split(oldString).length - 1
+      if (occurrences === 0)
+        throw new Error(
+          'writeMemory: `oldString` was not found in the target memory file. Call `readMemory` and copy it from ' +
+            'the current `content`; the attempted text may be stale or copied from non-memory session context.'
+        )
+      if (occurrences > 1)
+        throw new Error(
+          'writeMemory: `oldString` occurs multiple times — include more surrounding context to make it unique'
+        )
+      const updated = current.replace(oldString, newString)
+      return await deps.memory.write(scope, path, updated, undefined, 'tool')
+    }
+    const full = requireStringAllowEmpty(args, 'content')
+    return await deps.memory.write(scope, path, full, undefined, 'tool')
+  } catch (err) {
+    toToolError(err)
+  }
+}
+
+/** The record-memory surface for this call. External-memory tools are daemon-local but
+ *  operate on canonical records instead of pretending the plugin has files. The current
+ *  provider is re-resolved on EVERY call so a stale session tool cannot cross a provider or
+ *  capability change. The trusted agent scope is always ctx.agentId. */
+function recordSurface(ctx: SessionContext, deps: MemoryOpsDeps) {
+  const surface = deps.memory.adminSurfaceForAgent?.(ctx.agentId) ?? deps.memory.adminSurface()
+  if (!surface || surface.shape !== 'records') throw new Error('record memory is not available for this agent')
+  const scope = memoryScopeFor(ctx, deps)
+  const requireCapability = (operation: 'recall' | 'create' | 'get' | 'update' | 'delete'): void => {
+    if (!surface.capabilities.has(operation)) throw new Error(`record memory does not support ${operation}`)
+  }
+  return { surface, scope, requireCapability }
+}
+
+export async function searchMemory(
+  ctx: SessionContext,
+  args: Record<string, unknown>,
+  deps: MemoryOpsDeps
+): Promise<unknown> {
+  const { surface, scope, requireCapability } = recordSurface(ctx, deps)
+  requireCapability('recall')
+  const topK = optionalBoundedInt(args, 'topK', 1, 20) ?? 5
+  const maxBytes = optionalBoundedInt(args, 'maxBytes', 1, 32_768) ?? 8_192
+  const records = await surface.search(scope, {
+    turnId: randomUUID(),
+    query: requireString(args, 'query'),
+    topK,
+    maxBytes,
+    timeoutMs: 3_000
+  })
+  return { records }
+}
+
+export async function saveMemory(
+  ctx: SessionContext,
+  args: Record<string, unknown>,
+  deps: MemoryOpsDeps
+): Promise<unknown> {
+  const { surface, scope, requireCapability } = recordSurface(ctx, deps)
+  requireCapability('create')
+  const metadata = optionalObject(args, 'metadata')
+  const record = await surface.create(scope, {
+    operationId: randomUUID(),
+    text: requireString(args, 'text'),
+    ...(metadata ? { metadata } : {})
+  })
+  return { record }
+}
+
+export async function getMemory(
+  ctx: SessionContext,
+  args: Record<string, unknown>,
+  deps: MemoryOpsDeps
+): Promise<unknown> {
+  const { surface, scope, requireCapability } = recordSurface(ctx, deps)
+  requireCapability('get')
+  return { record: await surface.get(scope, requireString(args, 'id')) }
+}
+
+export async function updateMemory(
+  ctx: SessionContext,
+  args: Record<string, unknown>,
+  deps: MemoryOpsDeps
+): Promise<unknown> {
+  const { surface, scope, requireCapability } = recordSurface(ctx, deps)
+  requireCapability('update')
+  const metadata = optionalObject(args, 'metadata')
+  const version = optionalString(args, 'version')
+  const record = await surface.update(scope, {
+    operationId: randomUUID(),
+    id: requireString(args, 'id'),
+    text: requireString(args, 'text'),
+    ...(metadata ? { metadata } : {}),
+    ...(version ? { version } : {})
+  })
+  return { record }
+}
+
+export async function deleteMemory(
+  ctx: SessionContext,
+  args: Record<string, unknown>,
+  deps: MemoryOpsDeps
+): Promise<unknown> {
+  const { surface, scope, requireCapability } = recordSurface(ctx, deps)
+  requireCapability('delete')
+  const id = requireString(args, 'id')
+  const version = optionalString(args, 'version')
+  return {
+    id,
+    deleted: await surface.delete(scope, {
+      operationId: randomUUID(),
+      id,
+      ...(version ? { version } : {})
+    })
+  }
+}
